pyproject.toml
[tool.poetry]
name = "xai-shift-fingerprints"
version = "0.1.0"
description = "Attribution-driven dataset-shift diagnostics for CXR segmentation."
authors = ["Mohaisen Mohammed Lab <team@example.com>"]
readme = "README.md"
packages = [{ include = "xfp", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
numpy = "^1.25"
pandas = "^2.0"
scipy = "^1.11"
scikit-image = "^0.22"
scikit-learn = "^1.3"
pyyaml = "^6.0"
tqdm = "^4.66"
rich = "^13.7"
captum = "^0.7"
networkx = "^3.2"
plotly = "^5.20"
statsmodels = "^0.14"
pyarrow = "^15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
black = "^24.3"
ruff = "^0.4"
ipykernel = "^6.29"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
